New direction - avoid assignments inside 'while' and 'if' loops:
It appears that calling functions AND trying to assign to variables
from within (nested) if loops is a big no-no in python...in any case
it doesn't seem to work and trying leads to creating some very messy
code. Wil look into simply calling functions and making assignments in
the correct scope...Stack Overflow is a good friend!