Adds 'call_for_change' func. Adds gc to attempt reclaim of memory:
The 'call_for_change' function allows exit from the main loop and if statements
without passing a vriable parameter - this seems the only way to indirectly make
the calls necessary to recreate the color palette (as far as I can figure out!).
Any calls with variable parameters from within the main loop/ifs, fail to return
causing the prog to hang. Have tried the 'global' decloration from within the
loop and if statemnets - this caused no illumination of the neos at all.

The gc is leberaly placed in a bid to gain back vital memory. Seems the api
request is a major mem hog - this seems to have made a difference but not
enough to run adequately. There is a lot of 'stuttering' and auto-reset
occuring.
The new color is, however, being passed into the functions and the neos do
change color (once before freezing and resetting) - some success to a certain
degree :-)

I'm sure this 'leap-frogging' of functions and overuse of gc is very ugly and
probably concidered non pythonic.